Rename installed libraries to be more consistent with the usual Qt4/Qt5 lib naming conventions.
The built libs are now called libQtQuazip and libQt5Quazip on Qt4/Qt5,
respectively, in order to follow the usual Qt lib naming scheme. The
(cmake-generated) .pc file is also updated accordingly, and the
QuaZipConfig.cmake named according to the Qt version.
Additionally, the Version variable is fixed now in the .pc
pkconfig file.