Fix compiling when no Qt5Network and/or Qt5Test is available
CMake wants to properly configure qztest no matter what.
This could lead to trouble if QtNetwork and/or QtTest were
not available. This commit marks these as optional dependencies
and simply disables qztest altogether if those are not found.
This is especially important for Qt 6, where components not
asked for are not imported at all (see #84), leading to a failed build,
but this commit targets Qt 5 only for now.