Add .gitattributes
To avoid line endings mess, all files except NEWS.txt are now
forced to have LF line endings. They already have these, so
all this commit changes is prevent accidental line ending
changes in the future.

NEWS.txt was in CRLF to begin with, but that's harmless,
so it is left as it is. It doesn't make sense to correct
line endings now as it would break annotations (git blame).