Add symlink extraction test for Windows
While symlinks are generally not used on Windows
and require admin priviledges, we still need to
be able to extract archives with symlinks
(probably created on Linux).

This test simply checks that extraction works.
Somehow. It doesn't check the result, although
in the current version it seems that (broken)
shortcuts are created instead of symlinks.
They are probably harmless, but at least
other files are extracted successfully.