Update test platforms list in the docs
Add Astra Linux CE 1.6, these days the de-facto standard OS for Russian
government-funded organizations that don't deal with secret information
(there's Astra Linux SE for that, but it costs money).

It's also a useful platform to test on because it ships with Qt 5.11,
and that version wasn't being tested on any other platform before. Now
Qt 4.8.7, 5.9, 5.11, 5.12 and 5.15 are tested!