Clearly mark COPYING as the definitive license source
There are numerous copyright notices around the source tree. It's a
pain to maintain them and it messes up the history. The COPYING file is
there to contain all licensing information, but that could have been not
clear enough. Not it is explicitly stated in the COPYING file itself.
Not sure if that is enough legal-wise, but it certainly won't hurt.

Hope this fixes #101.