Add JlCompress extract to the root test
For Linux only: if the current directory is root,
and a file is being extracted into the current directory,
then the base extraction path is "/". In any other case
it doesn't end with an '/', so an additional '/' was added
in JlCompress to compose a common dir prefix. However,
if the path is "/" already, then adding another '/'
would yield "//", and then it's no longer a valid common
dir prefix, and the extraction will fail.

This was fixed in #132, but there was no test, so one was
added now. The fix was optimized a bit, so instead of appending
a '/' every time and then removing it back by calling QDir::cleanPath(),
a '/' is only added now if the prefix doesn't already end with a '/'.