fix deprecated QDateTime warning ctor with Qt 6.8
since Qt 6.8 constructor of QDateTime, which accepts TimeSpec enum,
marked as deprecated. to compile at this version of Qt it necessary to
explicitly pass QTimeZone instead.

to save work with lower version of Qt, version check macro was added.

it is critical to code where warnings treats as errors.

issue: #207