add large file unit tests
Creates a 5GB file with random data (in order to produce >4GB zip with high probability), then opens the zip and compares hashes with original content. This is a very expensive test so hid it behind an env TEST_ZIP_UNZIP_LARGE and only run it as part of alpine job. Expanded alpine job with Qt6.